feat : 상위 폴더 type 지정
