+++ src/types/folder.ts
@@ -1,0 +1,12 @@
+import { colorMap } from "@styles/colorMap";
+
+type Color = keyof typeof colorMap;
+
+export interface MainFolderItemProps {
+  folderId: number;
+  folderName: string;
+  createdAt: string;
+  noteCount: number;
+  folderColor: Color;
+  markState?: boolean;
+}
